Work around log4j cve
upstream log4j dependency was updated in #146 alas was not backported
yet to 3.x versions. This change removes its dependency altogether.
Since it is better practice to not depend on logging library directly
at all (but rather on a logging facade), and since upstream already
depends on a logging facade, we are better off.